Fix bug when adding achievements for groupmembers
